norm fix
