+++ Makefile
@@ -10,7 +10,7 @@ INCLUDE_DIR = ./include
 MAIN = main signal strarrutils
 CMD = cd echo env exec exec_proc exit export pwd unset
 PARSER =  escape_split expand_envvar expand_wildcard get_arg_num get_file_list get_length is load_lst parser rm_quotes
-REDIR = cmd_redir exec_redir utils_redir utils_redir2
+REDIR = cmd_redir exec_redir utils_redir utils_redir2 utils_redir3
 PIPE = utils_pipe utils_pipe2 utils_pipe3 utils_pipe4
 GET_NEXT_LINE = get_next_line_bonus get_next_line_utils_bonus
