+++ pipe/utils_pipe2.c
@@ -68,16 +68,13 @@ char	**repeat(char **args, char **envp, t_node *node)
 	}
 	else
 	{
-		if (node->redir_idx == 0)	
+		if (node->redir_idx == 0)
 			envp = find_command(args, envp, node);
 		if (node->redir_flag != 0)
 			backup_restor(node);
 		return (envp);
 	}
-	if (pid == 0)
-		exec_child(args, envp, node);
-	else
-		envp = exec_parents(pid, args, envp, node);
+	pipe_work(pid, args, envp, node);
 	return (envp);
 }
