+++ pipe/utils_pipe3.c
@@ -47,3 +47,11 @@ int	repeat_check(char **args, t_node *node)
 	}
 	return (0);
 }
+
+void	pipe_work(int pid, char **args, char **envp, t_node *node)
+{
+	if (pid == 0)
+		exec_child(args, envp, node);
+	else
+		envp = exec_parents(pid, args, envp, node);
+}
